clean up code #1
