+++ examples/mfmc/run_lofi_smc.py
@@ -12,11 +12,11 @@ from smcpy.mcmc.vector_mcmc_kernel import VectorMCMCKernel
 from smcpy import AdaptiveSampler as Sampler
 from smcpy.paths import GeometricPath
 
-from toy_helpers import M_LF, generate_noisy_data
+from exp_2d import M_LF as M_LF
+from exp_2d import generate_noisy_data
 from smcpy.mfmc_proposal import MultiFidelityProposal
-from smcpy.proposals import MultivarIndependent
 
-from toy_helpers import plot_2d_joint_posterior, plot_param_hists, plot_target_boxplots
+from plotting_helpers import plot_2d_joint_posterior, plot_param_hists, plot_target_boxplots
 
 # Data generation details
 STD_DEV = 0.2
@@ -32,7 +32,7 @@ Execute MF SMC
 '''
 
 # Setup low-fidelity case
-priors = [uniform(0.001, 2), uniform(0.5, 8)]
+priors = [uniform(0.01,2), uniform(0.05, 6)]
 vector_mcmc = VectorMCMC(M_LF, noisy_data, priors, STD_DEV)
 
 # initialize from prior
@@ -41,7 +41,7 @@ smc = Sampler(mcmc_kernel=mcmc_kernel, show_progress_bar=True)
 lofi_step_list, lofi_mll_list = smc.sample(
     num_particles=NUM_PARTICLES,
     num_mcmc_samples=5,
-    target_ess=0.9
+    target_ess=0.5
 )
 lofi_phi_list = smc.phi_sequence
 lofi_particles = lofi_step_list[-1].params
